clear hardware event listeners where appropriate.
This probably should have associated tests...but I'm not exactly sure what observable bad behavior
there'll be. Timers can be a bit messed up perhaps. The big observable one would be that if we
revert from established connection back to NoConnection, and then read an outgoing packet, we could
panic. But that transition isn't handled yet.