Specify IP packet on command line instead of inner packet length.
Generally use IP packet length as the canonical user-facing packet length. For example, if the user
specifies speed instead of a fixed packet interval, use the IP packet length to calculate the
interval (so the average IP bandwidth is as specified)

Some other cleanups to CLI options as well, and an incorrect PSK test I guess.